move collector to lib folder
