+++ lib/collector.js
@@ -32,7 +32,7 @@ const sampleSize = require('lodash/sampleSize');
 const uniqWith = require('lodash/uniqWith');
 const pickBy = require('lodash/pickBy');
 
-const { isFirstParty, getLocalStorage, safeJSONParse } = require('./lib/tools');
+const { isFirstParty, getLocalStorage, safeJSONParse } = require('./tools');
 
 async function run (uri, argv) {
   // const uri_ins = argv._[0];
@@ -63,13 +63,13 @@ async function run (uri, argv) {
   }
 
   // logger involves file access and should initate after overwrite check
-  const initLogger = require('./lib/logger');
+  const initLogger = require('./logger');
   const logger = initLogger(argv);
 
-  const { setup_cookie_recording } = require('./lib/setup-cookie-recording');
-  const { setup_beacon_recording } = require('./lib/setup-beacon-recording');
-  const { setup_websocket_recording } = require('./lib/setup-websocket-recording');
-  const { set_cookies } = require('./lib/set-cookies');
+  const { setup_cookie_recording } = require('./setup-cookie-recording');
+  const { setup_beacon_recording } = require('./setup-beacon-recording');
+  const { setup_websocket_recording } = require('./setup-websocket-recording');
+  const { set_cookies } = require('./set-cookies');
 
   const browser = await puppeteer.launch({
     headless: argv.headless,
@@ -99,7 +99,7 @@ async function run (uri, argv) {
     script: {
       host: os.hostname(),
       version: {
-        npm: require('./package.json').version,
+        npm: require('../package.json').version,
         commit: null,
       },
       cmd_args: process.argv.slice(2).join(' '),
@@ -317,7 +317,7 @@ async function run (uri, argv) {
   }
 
   // prepare regexp to match social media platforms
-  let social_platforms = yaml.safeLoad(fs.readFileSync(path.join(__dirname, 'assets/social-media-platforms.yml'), 'utf8')).map((platform) => {
+  let social_platforms = yaml.safeLoad(fs.readFileSync(path.join(__dirname, '../assets/social-media-platforms.yml'), 'utf8')).map((platform) => {
     return escapeRegExp(platform);
   });
   let social_platforms_regexp = new RegExp(`\\b(${social_platforms.join('|')})\\b`, 'i');
@@ -326,7 +326,7 @@ async function run (uri, argv) {
   });
 
   // prepare regexp to match links by their href or their caption
-  let keywords = yaml.safeLoad(fs.readFileSync(path.join(__dirname, 'assets/keywords.yml'), 'utf8')).map((keyword) => {
+  let keywords = yaml.safeLoad(fs.readFileSync(path.join(__dirname, '../assets/keywords.yml'), 'utf8')).map((keyword) => {
     return escapeRegExp(keyword);
   });
   let keywords_regexp = new RegExp(keywords.join('|'), 'i');
@@ -672,10 +672,10 @@ async function run (uri, argv) {
   }
 
   if (argv.output || argv.html) {
-    let html_template = argv['html-template'] || path.join(__dirname, 'assets/template.pug');
+    let html_template = argv['html-template'] || path.join(__dirname, '../assets/template.pug');
     let html_dump = pug.renderFile(html_template, Object.assign({}, output, {
       pretty: true,
-      basedir: __dirname,
+      basedir: path.join(__dirname, '..'),
       groupBy: groupBy,
       inlineCSS: fs.readFileSync(require.resolve('github-markdown-css/github-markdown.css')),
     }));
